(GH-540) Fix bolt upload failures
Prior to this commit we were seeing a lot of test failures when bolt was
trying to upload test files to the remote host. These files were
required to setup the env for the acceptance tests.

This commit fixes the issue by altering the parameters passed to
bolt_upload_file. In addition some common code has been refactored in to
spec_helper_acceptance_local to avoid duplication and drift.