support per-repo HTTP proxy for the git provider
Proxies can be specified either as a single proxy or as one or more
per-remote proxies.

Since the proxy config need not be supported for _every_ git command,
define a new wrapper function to be used for git operations that need
remote access, and then prepend the proxy arguments in that wrapper.

This implementation is better than the previously-documented method of
using proxies when:
a. Global configuration is not desired.
b. Initial `git clone` should/must use a proxy.

The previously-documented method is still available for different use
cases; reword the original instructions so that they are specified as an
alternative rather than the primary method.

Move all proxy information into its own section, since this has become
rather large to be included in general git usage.