(MODULES-11335) README update
Prior to this commit, cleaning local changes on agent run was a feature
that could be assumed to be available in the modules functionality by
the name of some of the parameters, when in reality, this behavior is
not possible.

This commit aims to update the README to include a couple of notes
clarifying the current limitations of the module.